Do not enqueue promises, enqueue async functions.
We now want the ability to wait to invoke the promise generating
function so that we can delay running synchronous operations until we're
ready to handle their rejections. This is for when the operation may
have a hot cache and no asynchronous operations are necessary, but if
they are we want to wait for the user to get around to calling an
`async` method on our interface.

Closes #6.